+++ tests/SonarqubeInstanceTest.php
@@ -5,6 +5,7 @@ namespace Tests\ForgeQC\SonarqubeApiClient;
 use PHPUnit\Framework\TestCase;
 use ForgeQC\SonarqubeApiClient\HttpClient;
 use ForgeQC\SonarqubeApiClient\SonarqubeInstance;
+use ForgeQC\SonarqubeApiClient\SonarqubeProject;
 use GuzzleHttp\Exception\ClientException;
 use Dotenv\Dotenv;
 
@@ -35,6 +36,66 @@ class SonarqubeInstanceTest extends TestCase
     $this->assertArrayHasKey('name', $projects[0]);
   }
 
+  //Test getProjects() function on sonarqube online instance
+  public function testGetMultipleProjectsMeasures()
+  {
+    //Connect to sonarqube API
+    $api = new HttpClient('https://sonarcloud.io/api/');
+    $instance = new SonarqubeInstance($api);
+
+    $measures = $instance->getMultipleProjectsMeasures('Board-Voting,paysuper_paysuper-currencies');
+
+    $this->assertArrayHasKey('Board-Voting', $measures);
+    $this->assertArrayHasKey('paysuper_paysuper-currencies', $measures);
+    $this->assertArrayHasKey('sqale_rating', $measures['Board-Voting']);
+    $this->assertArrayHasKey('bugs', $measures['Board-Voting']);
+    $this->assertArrayHasKey('reliability_remediation_effort', $measures['Board-Voting']);
+    $this->assertArrayHasKey('security_rating', $measures['Board-Voting']);
+    $this->assertArrayHasKey('vulnerabilities', $measures['Board-Voting']);
+    $this->assertArrayHasKey('sqale_rating', $measures['Board-Voting']);
+    $this->assertArrayHasKey('security_remediation_effort', $measures['Board-Voting']);
+    $this->assertArrayHasKey('coverage', $measures['Board-Voting']);
+
+    $measuresCustom = $instance->getMultipleProjectsMeasures('Board-Voting,paysuper_paysuper-currencies','sqale_index');
+    $this->assertArrayHasKey('Board-Voting', $measuresCustom);
+    $this->assertArrayHasKey('sqale_index', $measuresCustom['Board-Voting']);
+  }
+
+  function testAggregateMultipleProjectsMeasures()
+  {
+    //Connect to sonarqube API
+    $api = new HttpClient('https://sonarcloud.io/api/');
+    $instance = new SonarqubeInstance($api);
+
+    $project1 = new SonarqubeProject($api, 'Board-Voting');
+    $measures1 = $project1->getMeasures();
+
+    $project2 = new SonarqubeProject($api, 'paysuper_paysuper-currencies');
+    $measures2 = $project1->getMeasures();
+
+    $aggregatedMeasures = $instance->aggregateMultipleProjectsMeasures('Board-Voting,paysuper_paysuper-currencies');
+
+    $this->assertSame(2, $aggregatedMeasures['projects_count_request']);
+    $this->assertSame(2, $aggregatedMeasures['projects_count_with_measures']);
+    $this->assertSame(round(($measures1['reliability_rating']+$measures2['reliability_rating'])/2, 0, PHP_ROUND_HALF_UP),$aggregatedMeasures['reliability_rating']);
+    $this->assertSame(round(($measures1['sqale_rating']+$measures2['sqale_rating'])/2, 0, PHP_ROUND_HALF_UP),$aggregatedMeasures['sqale_rating']);
+    $this->assertSame(round(($measures1['security_rating']+$measures2['security_rating'])/2, 0, PHP_ROUND_HALF_UP),$aggregatedMeasures['security_rating']);
+
+
+    //Test with one project non existing in Sonarqube
+    $aggregatedMeasuresHalfExists = $instance->aggregateMultipleProjectsMeasures('Board-Voting,non-existing-project');
+    $this->assertSame(2, $aggregatedMeasuresHalfExists['projects_count_request']);
+    $this->assertSame(1, $aggregatedMeasuresHalfExists['projects_count_with_measures']);
+    $this->assertSame(round($measures1['reliability_rating'], 0, PHP_ROUND_HALF_UP),$aggregatedMeasuresHalfExists['reliability_rating']);
+    $this->assertSame(round($measures1['sqale_rating'], 0, PHP_ROUND_HALF_UP),$aggregatedMeasuresHalfExists['sqale_rating']);
+    $this->assertSame(round($measures1['security_rating'], 0, PHP_ROUND_HALF_UP),$aggregatedMeasuresHalfExists['security_rating']);
+
+    //Test with no projects existing in sonarqube. Should return an empty array.
+    $aggregatedMeasuresNoneExists = $instance->aggregateMultipleProjectsMeasures('non-existing-projectA,non-existing-projectB');
+    $this->assertSame([], $aggregatedMeasuresNoneExists);
+
+  }
+
   //Test createGroup() function
   public function testCreateDeleteGroup()
   {
@@ -55,7 +116,18 @@ class SonarqubeInstanceTest extends TestCase
 
   }
 
-  //Test createGroup() function
+  //Test userExists() function
+  public function testUserExists()
+  {
+    //Connect to sonarqube API
+    $api = new HttpClient('https://next.sonarqube.com/sonarqube/api/');
+    $instance = new SonarqubeInstance($api);
+
+    $this->assertSame(true, $instance->userExists('admin'));
+    $this->assertSame(false, $instance->userExists('non-existing-user'));
+  }
+
+  //Test createUser() function
   public function testCreateDeleteUser()
   {
     //Tel PHPUNIT that the correct behavior of the tested function is to throw an exception
@@ -67,8 +139,8 @@ class SonarqubeInstanceTest extends TestCase
     $api = new HttpClient('https://sonarcloud.io/api/', $sonar_api_key);
     $instance = new SonarqubeInstance($api, 'testapi');
 
-    $user = $instance->createUser('TestUser', 'TestUser', 'test@user.local');
-
+    $user = $instance->createUser('jdoe', 'John DOE', 'test@user.local');
+    $userUpdated = $instance->updateUser('jdoe', 'John DOE Updated', 'test-updated@user.local');
   }
 
 }